+++ src/netedit/changes/GNEChange_MeanData.cpp
@@ -1,6 +1,6 @@
 /****************************************************************************/
 // Eclipse SUMO, Simulation of Urban MObility; see https://eclipse.dev/sumo
-// Copyright (C) 2001-2024 German Aerospace Center (DLR) and others.
+// Copyright (C) 2001-2025 German Aerospace Center (DLR) and others.
 // This program and the accompanying materials are made available under the
 // terms of the Eclipse Public License 2.0 which is available at
 // https://www.eclipse.org/legal/epl-2.0/
@@ -24,6 +24,9 @@
 #include <config.h>
 
 #include <netedit/GNENet.h>
+#include <netedit/GNEViewNet.h>
+#include <netedit/GNEViewParent.h>
+#include <netedit/GNEApplicationWindow.h>
 #include <netedit/elements/data/GNEMeanData.h>
 
 #include "GNEChange_MeanData.h"
@@ -45,17 +48,18 @@ GNEChange_MeanData::GNEChange_MeanData(GNEMeanData* meanData, bool forward) :
 
 
 GNEChange_MeanData::~GNEChange_MeanData() {
-    myMeanData->decRef("GNEChange_MeanData");
-    if (myMeanData->unreferenced()) {
-        // show extra information for tests
-        WRITE_DEBUG("Deleting unreferenced " + myMeanData->getTagStr());
-        // make sure that MeanData isn't in net before removing
-        if (myMeanData->getNet()->getAttributeCarriers()->retrieveMeanData(myMeanData->getTagProperty().getTag(), myMeanData->getID(), false)) {
-            // delete MeanData from net
-            myMeanData->getNet()->getAttributeCarriers()->deleteMeanData(myMeanData);
+    // only continue we have undo-redo mode enabled
+    if (myMeanData->getNet()->getViewNet()->getViewParent()->getGNEAppWindows()->isUndoRedoAllowed()) {
+        myMeanData->decRef("GNEChange_MeanData");
+        if (myMeanData->unreferenced()) {
+            // make sure that MeanData isn't in net before removing
+            if (myMeanData->getNet()->getAttributeCarriers()->retrieveMeanData(myMeanData->getTagProperty().getTag(), myMeanData->getID(), false)) {
+                // delete MeanData from net
+                myMeanData->getNet()->getAttributeCarriers()->deleteMeanData(myMeanData);
+            }
+            // delete mean data
+            delete myMeanData;
         }
-        // delete mean data
-        delete myMeanData;
     }
 }
 
@@ -63,8 +67,6 @@ GNEChange_MeanData::~GNEChange_MeanData() {
 void
 GNEChange_MeanData::undo() {
     if (myForward) {
-        // show extra information for tests
-        WRITE_DEBUG("Removing " + myMeanData->getTagStr() + " '" + myMeanData->getID() + "' in GNEChange_MeanData");
         // unselect if mySelectedElement is enabled
         if (mySelectedElement) {
             myMeanData->unselectAttributeCarrier();
@@ -72,8 +74,6 @@ GNEChange_MeanData::undo() {
         // delete meanData from net
         myMeanData->getNet()->getAttributeCarriers()->deleteMeanData(myMeanData);
     } else {
-        // show extra information for tests
-        WRITE_DEBUG("Adding " + myMeanData->getTagStr() + " '" + myMeanData->getID() + "' in GNEChange_MeanData");
         // select if mySelectedElement is enabled
         if (mySelectedElement) {
             myMeanData->selectAttributeCarrier();
@@ -89,8 +89,6 @@ GNEChange_MeanData::undo() {
 void
 GNEChange_MeanData::redo() {
     if (myForward) {
-        // show extra information for tests
-        WRITE_DEBUG("Adding " + myMeanData->getTagStr() + " '" + myMeanData->getID() + "' in GNEChange_MeanData");
         // select if mySelectedElement is enabled
         if (mySelectedElement) {
             myMeanData->selectAttributeCarrier();
@@ -98,8 +96,6 @@ GNEChange_MeanData::redo() {
         // insert meanData into net
         myMeanData->getNet()->getAttributeCarriers()->insertMeanData(myMeanData);
     } else {
-        // show extra information for tests
-        WRITE_DEBUG("Removing " + myMeanData->getTagStr() + " '" + myMeanData->getID() + "' in GNEChange_MeanData");
         // unselect if mySelectedElement is enabled
         if (mySelectedElement) {
             myMeanData->unselectAttributeCarrier();
